rm dead code branch.
courtesy of PHPStan lvl 5.
explanation: that local variable was never instantiated as a Reference
object, thus remaining NULL throughout parsing, thus the if conditional
was always evaluating to FALSE.
turns out that we don't need to local Referencevariable after all - we want to set concept
id (UI) and name directly on the Concept object. which we've been doing.
in sum - not a bugfix, just cleanup.