Updates MeSH parser for 2025 release.
the DTD for descriptors changed mid-2024, and replaced the single
RegistryNumber element in Concept with an optional
RegistryNumberList element that can contain multiple RegistryNumber
elements.
We don't need fork logic here since this change already went into
effect, and we're only supporting the latest 2024 DTD and the 2025 DTD.
In other words, a simple drop-in replacement is all we need here.